feat: setting chalk_log_level to trace in connect.c4m (#365)
This will help with debugging as chalk report will include all trace
logs as part of the report in `_OP_ERRORS` field. In the future might be
worth renaming the field as `ERRORS` sounds dangerous even when it
contains trace-level logs...

We should do a pass on anything potentially sensitive we might be
logging to ensure the field does not leak anything security-sensitive.