refactor(cloudMetadata): split metadata collection into host-specific procs
There was previously one long function that handled collecting metadata
for AWS, Azure, and GCP, which had to return after setting metadata for
one platform. Split into separate functions, and add an enum to resolve
the FIXMEs.

We can refactor further later, but this is a quick step in the right
direction.