refactor(cloudMetadata): refactor metadata collection (#367)
* refactor(cloudMetadata): don't qualify toLowerAscii() call

Use the standard style in Nim: omit the module name for an imported
function unless ambiguous. This allows use of the method-call syntax.

For background, see e.g. Imports in Nim [1]

[1] https://narimiran.github.io/2019/07/01/nim-import.html

* refactor(cloudMetadata): write startsWith consistently

* refactor(cloudMetadata): change snake_case identifier to camelCase

Be consistent with the rest of the codebase.

* style(cloudMetadata): remove some newlines

* refactor(cloudMetadata): split metadata collection into host-specific procs

There was previously one long function that handled collecting metadata
for AWS, Azure, and GCP, which had to return after setting metadata for
one platform. Split into separate functions, and add an enum to resolve
the FIXMEs.

We can refactor further later, but this is a quick step in the right
direction.

* refactor(cloudMetadata): move AWS-specific procs/templates down

Communicate to the reader that the metadata collection for Azure and GCP
does not require these routines.

* refactor(cloudMetadata): declare two variables via `let` section

* refactor(cloudMetadata): avoid a variable declaration

Communicate that this information is only needed for the case statement.

Also improve formatting.