refactor(cloudMetadata): turn templates into procs (#368)
Commit 55cd0d520352 (2024-07-09, "fix(cloudMetadata): set
_OP_CLOUD_PROVIDER earlier") reminded me that we should make this file
easier to reason about. Here's another quick improvement.

Previously, at a call site for e.g. `oneItem` it was not obvious that:

- It mutates `result`.
- It uses `token`.

Replace the templates with procs to make that clearer, and add doc
comments. This also removes the excessive inlining.

From the Nim Manual [1]:

    Style note: For code readability, it is best to use the least
    powerful programming construct that remains expressive. So the
    "check list" is:

    1. Use an ordinary proc/iterator, if possible.
    2. Else: Use a generic proc/iterator, if possible.
    3. Else: Use a template, if possible.
    4. Else: Use a macro.

Also make a few minor formatting changes.

[1] https://nim-lang.org/2.0.0/manual.html#macros