refactor: reduce AttrScope usage, toward fully-dotted keys (#337)
Continue to prepare for the new con4m, by refactoring so that we get/set
everything using the top-level chalk scope with fully-dotted keys. That
helps prepare for the new con4m lookup interface, which in con4m v2 was:

    proc lookup*[T](ctx: RuntimeState, key: string): Option[T] =

where `key` is fully-dotted. And similarly in libcon4m:

    extern c4m_value_t *
    c4m_vm_attr_get(c4m_vmthread_t *tstate,
                    c4m_str_t      *key,
                    bool           *found);

Previously we were passing subscopes around, but `AttrScope` doesn't
exist in the new con4m (where there's essentially one big dictionary of
fully-dotted keys).

We will refactor this further later.

With this commit, the only remaining `AttrScope`s are:

    $ git grep --ignore-case --break --heading AttrScope
    src/run_management.nim
    20:proc getChalkScope*(): AttrScope =
    27:    if v.isA(AttrScope):
    30:proc sectionExists*(scope: AttrScope, s: string): bool =
    41:proc con4mAttrSet*(attrs: AttrScope, fqn: string, value: Box, attrType: Con4mType) =
    
    src/sinks.nim
    177:proc getAuthConfigByName*(name: string, attr: AttrScope = AttrScope(nil)): Option[AuthConfig] =

Refs: https://github.com/crashappsec/chalk/issues/269