feat: add libcon4m ABI wrapper (updated for 2024-07-25)
Add the lower-level Nim wrapper for libcon4m, updated for the latest
libcon4m `main` at the time of writing. I've tidied this wrapper a
little - I don't think more is worthwhile for now.

My preferred/usual pattern is to have two separate wrappers:

- One for the low-level interface, where it's sometimes easier to just
  wrap everything, especially if the underlying library is still
  receiving somewhat-frequent updates.

- One for a higher-level interface that's more ergonomic/idiomatic,
  where we'll wrap only what we need.

Note that most procs are annotated with the custom `lc4m` pragma, which
is defined as:

    {.pragma: lc4m, cdecl, importc.}

so the name for most procs is the same as the C function it wraps.
However, from Nim's naming rules for identifiers [1]:

    Identifiers in Nim can be any string of letters, digits and
    underscores, with the following restrictions:

    - begins with a letter
    - does not end with an underscore _
    - two immediate following underscores __ are not allowed

a proc cannot begin with an underscore. Therefore, we prefix with
`internal` the proc name for any C function name that begins with an
underscore. For example:

    proc internal_c4m_print*(a0: c4m_obj_t): void {.cdecl, varargs, importc: "_c4m_print".}

Note to self: double check some of the more recent changes to
meson.build (the conditional compilation that affects the wrapper).

[1] https://nim-lang.org/2.0.8/manual.html#lexical-analysis-identifiers-amp-keywords