test: fixing integration tests
changed semantic of dockerfile param in docker build wrappers
which broke some --slow tests so adding new param content now.
In theory we could rename dockerfile to dockerfile_path and dockerfile
params but that would require changing all existing callers which is
bigger refactor we can opt in to do in the future. For now content param
should suffice.