refactor: much faster docker image extract (#404)
previously we were trying:

* `docker run --entrypoint=cat <image> /chalk.json`
* `docker save <image> && unzip ...`

This was pretty fast for images which had `cat` available however
it was much slower for distroless images. For example chalks own image
`crashappsec/chalk` was pretty slow to extract as its distroless.
In addition `cat` only worked if docker was able to execute `cat`
with platform compatibility.

Right now we simply create a container but we never start it.
It then allows to extract a file from the container via `docker cp`
without too much fuss similarly how we were extracting it before.
As the container is never started its version compatibility is
irrelevant as nothing actually runs. Then we simply remove container.
Simple and fast.