feat: adding pingttl util
This will allow to get transient IP addresses along the NAT route
so that we can fetch intermediate IPs from within docker containers
which might not be accessible otherwise. For example it can be used
to get the subnet gateway IP which is not otherwise available
as within the docker container unless its running with host networking
chalk can only get the docker network private IP address.
This should be useful in figuring out where within the network
container is running.