fix: disabling heartbeats in lambda by default
In lambda we cannot control its runtime and as such the heartbeat rate
cannot be honored so we disable it altogether.
Instead lambda "liveliness" can be defined by the lambda deployment
itself vs relying on runtime heartbeats. For example a lambda is still
"deployed" even if its never triggered and so its runtime will never
send any heartbeats.