fix: embed attestation can lookup chalk keys in multiple paths
Previously embed provider only looked up keys in a single location
which by default was current working directly.
This is not desirable in CI as placing additional files in PWD
can have other side-effects.
As such setup-chalk-action was placing the generated key relative
to chalk binary however chalk was not able to lookup that key.