fix: parsing git timestamp offsets
When git reports its timestamp, its localized with the provided offset
and as chalk wasnt accounting for that, in some cases, commit timestamps
would show up as future timestamps which is very confusing

Now all timestamps are normalized with the provided offset.

In addition all date reporting was changed to use consistent output
format of ISO-8601.