fix: context manager to scope logs collection for _OP_ERRORS and ERR_INFO
without context manager it was a manual process which was setting the
current error object which is responsible for populating either
_OP_ERRORS or ERR_INFO keys. As such for example during docker build,
as soon as chalk was starting subscan of the docker binary, its target
was switched to the docker binary chalk object hence no more logs were
collected for `_OP_ERRORS` even after subscan was complete.

Now operations explicitly have to use context manager to limit the scope
of the current error object hence guaranteeing scope is always reverted
to the original object. Docker collection is more involved due to
multi-platform builds as a single base-chalk is used to create possibly
many subplatform chalk objects hence the context manager is necessary
in a few surgical places.