fix: empty buildx node config (#462)
not all build nodes have a config file associated with them and as such
buildkitd.toml might not exist. That doesnt mean the node doesnt exist,
it simply means it has no custom config.

For example:

```
docker buildx create test
```

To correctly detect whether a node has a config or not, the container
entrypoint/cmd are inspected to find the corresponding `--config`
and if present, only then attempts to read the node config.
Otherwise empty config is used.

This also surfaced a segfault when using empty config the code was
assuming ["registry"] was always present which is not the case.

As the node code now needed to inspect the container args, moved
node-related logic to separate nodes.nim therefore allowing it to
reference inspect.nim as otherwise would result in circular import
as inspect.nim needs exe.nim to make docker subprocess calls.