fix: _REPO_TAGS missing all pushed tags
In order to reduce registry interactions, only unique repo digests are
interacted with however that meant that if the same image was tagged
with multiple tags, its tags were lost as the locally loaded RepoTags
will be missing all other tags. The fix is simply concatenate all given
tags when processing tags.