fix: check buildkit frontend version to determine context arg compatibility
Not all older frontends support `--build-arg` and sometimes they do no
specify the version as part of the syntax directive. For example:

```
syntax: docker/dockerfile:experimental
```

Chalk cannot infer anything about the meaning of `experimental`
as its not pinned to any version. Therefore we call the dockerfile
frontend container and check its actual version.
This is possibly not as efficient even if the version is specified
however that would also require some heuristics. For example:

```
syntax: docker/dockerfile:1
```

As `1` can change over time it might not be clear which syntax should
actually be honored. Therefore it is probably a lot simpler to simply
ask it what version it is instead of possibly error-prone version
detection heuristics.