feat: initial cert plugin (#515)
* feat: initial cert plugin

* feat: codec search() to return seq[ChalkObj]

as a single cert file can have multiple cert artifacts, we need a way
to accommodate that as its no longer 1-1 everywhere

* refactor: collecting all hashes (unchalked/pre/ending) in collect

previously some hashes were collected as part of the collect logic and
others were computed in system plugin which was confusing.

* feat: add getOsFileHandle to fd_cache

required some shenanigans as File object is internal to FileStream

* wip certs segfault

* Bug fix

* fix: cleanup after segfault fix

* fix: segfaults and other edge cases - many symlink related

* feat: add symlink_behavior_[non]_chalking configs

this allows to split behavior between chalking and non-chalking commands

useful for certs as most certs are usually symlinked and so we do not
want to skip them by default as thats the default chalking behavior

* build: bump con4m and clean up Dockerfile how deps are installed

* fix: not running runtime collecting before chalking artifacts during insert

* fix: dont skip codecs collection, regardless of resourceTypes

* test: fix container id hash matching

* docs: cert keys and adding more keys

* refactor: scanArtifactLocations to iterate over FS once

previosly the outer loop was by the codec hence FS was scanned once per codec
which is very inefficient for large scans such as scanning / as:

* fd_cache cannot keep all the FDs open at the same time
* OS cant keep all the IO in-cache hence scanning files might result
  more disk-interactions per file

now the outer loop is the FS walk and the for each path candidate, all
codecs are consulted to find chalkmarks in them

* feat: support multiple certs in single file

* refactor: allow to pick codecs for scanning

* fix: ELF codec handle .o files

.o files might not have program table and so we should skip them

also only scanning supported elf types as we do not want to scan things
like coredump

* fix: pyc codec returned some(nil) in some cases

* fix: after codec found chalkmark, skip all other codecs

otherwise same file can be scanned by multiple codecs which is not
expected. for example source codec is meant as a fallback in some cases
and so if earlier codec found a valid chalkmark, source codec might find
additional dummy chalkmark

* feat: always rpoert _OP_ARTIFACT_PATH even during non-chalking op

* fix: dedup chalk objects by chalk id

otherwise we get duplicate chalkmarks for example when the file is
copied on the FS

* feat: restrict docker subscan to specific codecs

by default only certs are only looked for

* refactor: honor ignore_patterns in all commands

* build: bump con4m

* test: cert test extract test

* refactor: split util.nim to utils/*

there were couple of reasons:

* chalk had primary imported module config.nim which had dep chain:

  config.nim -> run_management.nim -> chalk_common.nim

  This allowed to import a single module in most of the chalk codebase
  which will automatically provide most stdlib deps.

  That however caused some issues due to:

* util.nim was dependant on config.nim hence it was impossible to use
  any of the utils in the config.nim without adding circular dependency
  without any ugly hacks like defining function stub...

* util.nim was a pretty big helper file which was getting hard to
  maintain

By splitting all of the utils in much smaller modules it allows to
reference subset of primitive utils (like strings.nim) anywhere in chalk
without any circular deps nonsense

In addition this commit removes a bunch of magical imports config.nim
was providing before in favor of explicit imports for most things
which is more intuitive and explicit where methods come from
without needing to know what are all exported deps in chalk_common.nim.

Finally chalk_common.py was mostly defining base types without much more
other logic hence renamed it to types.nim.

* cleanup: remove getlibpath

* fix: certs chaching chalkmark for future validation

otherwise if there are differences in mark/report templates
METADATA_ID validation can fail

* fix: unit tests imports

* test: add docker cert subscan test case

* feat: _OP_ARTIFACT_PATH_WITHIN_VCTL for the certs path in git

* feat: hacky env var certs collection

this is not the final approach but a poc of the feature

* feat: collecting env vars

instead of an ugly hack in certs codec, adding collection mechanism
to collect artifacts from env vars

* build: bump con4m with install headers fix for mac

* docs: document env_vars config in CHANGELOG

* feat: handle base64 encoded env var certs

* docs: better comment explaining why files are looped over first

---------

Co-authored-by: John Viega <john@crashoverride.com>