feat: github job id
now that job.check_run_id is available, use it to construct
full github build uri

previously only including the workflow id in the build uri had some edge
cases:

* matrix builds. It was not possible to determine from which job did the
  chalk report originate from as multiple jobs were grouped under a
  single workflow
* multiple named jobs in a single workflow. Similarly all jobs share the
  same workflow id hence the build URI wasnt precise enough

This was a GitHub limitation which is now partially resolved as it
exposes the job id as a github context variable although not yet as an
env var therefore for now chalk will rely on setup-chalk-action actually
setting the env var for chalk to read. Otherwise it will fallback to
previous behavior.